% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_clean}
\alias{db_clean}
\title{Clean all tables and definition}
\usage{
db_clean(con)
}
\arguments{
\item{con}{an existing database connection}
}
\value{
No return
}
\description{
Clean all tables and definition
}
