% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{Disconnect to db}
\usage{
db_disconnect(con)
}
\arguments{
\item{con}{a connection object as produced by dbConnect.}
}
\value{
no return values
}
\description{
Disconnect to db
}
\examples{
\dontrun{
con <- db_disconnect()
dbDisconnect(con)
}
}
