% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_absolute_path}
\alias{.check_absolute_path}
\title{Check absolute path for system}
\usage{
.check_absolute_path(path)
}
\arguments{
\item{path}{File path to check}
}
\value{
No return is expected
}
\description{
Check absolute path for system
}
