% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{.run_slurm_cmd}
\alias{.run_slurm_cmd}
\title{Execute a Slurm command locally or remotely}
\usage{
.run_slurm_cmd(cmd, host, username = NULL)
}
\description{
Only allows Slurm commands: sbatch, squeue, scancel
}
