% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_list}
\alias{project_list}
\title{List all projects}
\usage{
project_list(con = NULL)
}
\arguments{
\item{con}{a db connection}
}
\value{
a table for all projects
}
\description{
List all projects
}
