% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_resource_get}
\alias{project_resource_get}
\title{Get resources of a project}
\usage{
project_resource_get(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\value{
a table of resources used in the project
}
\description{
Get resources of a project
}
