% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_resource.R
\name{project_resource_log_delete}
\alias{project_resource_log_delete}
\title{Delete all logs for a project resource}
\usage{
project_resource_log_delete(project, resource, con = NULL)
}
\arguments{
\item{project}{project name}

\item{resource}{resource name}

\item{con}{connection to database}
}
\value{
no return
}
\description{
Delete all logs for a project resource
}
