% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_start}
\alias{project_start}
\title{Start a project}
\usage{
project_start(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\value{
no return
}
\description{
Start a project
}
