% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_status}
\alias{project_status}
\title{Get project status}
\usage{
project_status(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{a connection to database}
}
\value{
no return
}
\description{
Get project status
}
