% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_stop}
\alias{project_stop}
\title{Stop a project}
\usage{
project_stop(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\value{
no return
}
\description{
Stop a project
}
