% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{schedule_options}
\alias{schedule_options}
\title{Set or get options for my package}
\usage{
schedule_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{sensor_url}}{ The url for sensor API}
}
}

