% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{table_exist}
\alias{table_exist}
\title{Check whether a table is existed}
\usage{
table_exist(table, con = NULL)
}
\arguments{
\item{table}{table name}

\item{con}{a connection}
}
\value{
logical value
}
\description{
Check whether a table is existed
}
