% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_clean}
\alias{task_clean}
\title{Clear all tasks in a project}
\usage{
task_clean(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{a connection to database}
}
\value{
no return
}
\description{
Clear all tasks in a project
}
