% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_status}
\alias{task_status}
\title{Get task status in a project}
\usage{
task_status(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{a connection to database}
}
\value{
a table to show task status
}
\description{
Get task status in a project
}
