% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{taskqueue_options}
\alias{taskqueue_options}
\title{Set or get options for my package}
\usage{
taskqueue_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
Options for task queue
}
\description{
Set or get options for my package
}
\section{Supported options}{

The following options are supported
\itemize{
 \item{\code{host}}{host of postgreSQL database}
 \item{\code{port}}{port of postgreSQL database}
 \item{\code{user}}{user name of postgreSQL database}
 \item{\code{password}}{pasword of postgreSQL database}
 \item{\code{database}}{database name of postgreSQL database}
}
}

