% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_connect}
\alias{db_connect}
\title{Connect to postgreSQL db}
\usage{
db_connect(con = NULL)
}
\arguments{
\item{con}{An existing connection. Return the existing connection if not null}
}
\value{
a connection object as produced by dbConnect
}
\description{
Connect to postgreSQL db
}
\examples{
\dontrun{
con <- db_connect()
}
}
