% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_add}
\alias{project_add}
\title{Create project to database}
\usage{
project_add(project, table = paste("task", project, sep = "_"), ...)
}
\arguments{
\item{project}{project name}

\item{table}{A table name for this project.}

\item{...}{Other arguments for project setting}
}
\description{
Create project to database
}
