% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_add}
\alias{project_add}
\title{Create project to database}
\usage{
project_add(project, ...)
}
\arguments{
\item{project}{project name}

\item{...}{Other arguments for project setting}
}
\description{
Create project to database
}
