% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_add}
\alias{project_add}
\title{Create project to database}
\usage{
project_add(project, memory = 10)
}
\arguments{
\item{project}{project name}

\item{memory}{memory usage in GB}
}
\value{
no returns
}
\description{
Create project to database
}
