% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_delete}
\alias{project_delete}
\title{Delete a project}
\usage{
project_delete(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\description{
Delete a project
}
