% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_get}
\alias{project_get}
\title{Get a project}
\usage{
project_get(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\description{
Get a project
}
