% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_reset}
\alias{project_reset}
\title{Reset a project}
\usage{
project_reset(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\value{
no return
}
\description{
Reset a project
}
