% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_reset}
\alias{project_reset}
\title{Reset status of all tasks in a project to NULL}
\usage{
project_reset(project, con = NULL)
}
\arguments{
\item{project}{project name}

\item{con}{connection to database}
}
\description{
Reset status of all tasks in a project to NULL
}
