% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_reset}
\alias{project_reset}
\title{Reset a project}
\usage{
project_reset(project, log_clean = TRUE)
}
\arguments{
\item{project}{project name}

\item{log_clean}{whether to clean log files}
}
\value{
no return
}
\description{
Reset a project
}
