% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_reset}
\alias{project_reset}
\title{Reset status of all tasks in a project to NULL}
\usage{
project_reset(project, status = c("working", "failed"), con = NULL)
}
\arguments{
\item{project}{project name}

\item{status}{status to reset (e.g. working, failed, or all), all tasks if status = all}

\item{con}{connection to database}
}
\value{
no return
}
\description{
Reset status of all tasks in a project to NULL
}
