% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_resource.R
\name{project_resource_add}
\alias{project_resource_add}
\title{Assign a resouce to a project}
\usage{
project_resource_add(
  project,
  resource,
  working_dir,
  account = NULL,
  times = 1,
  workers = NULL
)
}
\arguments{
\item{project}{project name}

\item{resource}{resource name}

\item{working_dir}{working directory for this resource}

\item{account}{optional for account to use this resource}

\item{times}{worker wall times for this resource}

\item{workers}{maximum workers for this project.}
}
\description{
Assign a resouce to a project
}
