% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_resource.R
\name{project_resource_add_jobs}
\alias{project_resource_add_jobs}
\title{Add a job to project/resource}
\usage{
project_resource_add_jobs(project, resource, job, reset = FALSE)
}
\arguments{
\item{project}{project name}

\item{resource}{resource name}

\item{job}{job name. If missing, reset to empty}

\item{reset}{whether to reset jobs}
}
\value{
no return
}
\description{
Add a job to project/resource
}
