% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource_add}
\alias{resource_add}
\title{Add a new resource}
\usage{
resource_add(
  name,
  type = c("slurm", "computer"),
  host,
  workers,
  log_folder,
  nodename = strsplit(host, "\\\\.")[[1]][1],
  con = NULL
)
}
\arguments{
\item{name}{resource name}

\item{type}{resource type (e.g. slurm or computer)}

\item{host}{host name}

\item{workers}{worker number}

\item{log_folder}{log folder which has to be absolute path.}

\item{nodename}{nodename obtained by Sys.info()}

\item{con}{a db connection}
}
\value{
no return
}
\description{
Add a new resource
}
