% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{resource_get}
\alias{resource_get}
\title{Get a resource}
\usage{
resource_get(resource, con = NULL)
}
\arguments{
\item{resource}{resource name}

\item{con}{a db connection}
}
\value{
a data frame for selected resource
}
\description{
Get a resource
}
