% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_add}
\alias{task_add}
\title{Add tasks into a project}
\usage{
task_add(project, num, clean = TRUE, con = NULL)
}
\arguments{
\item{project}{project name}

\item{num}{number of tasks}

\item{clean}{whether to clean existing tasks}

\item{con}{a connection to database}
}
\description{
Add tasks into a project
}
