% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{task_get}
\alias{task_get}
\title{Get tasks by status}
\usage{
task_get(project, status = c("failed"), limit = 10, con = NULL)
}
\arguments{
\item{project}{project name}

\item{status}{status to reset (e.g. working, failed, or all), all tasks if status = all}

\item{limit}{number of rows to return}

\item{con}{connection to database}
}
\value{
A data frame of tasks
}
\description{
Get tasks by status
}
