% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{taskqueue_options}
\alias{taskqueue_options}
\title{Set or get options for my package}
\usage{
taskqueue_options(...)
}
\arguments{
\item{...}{Option names to retrieve option values or \code{[key]=[value]} pairs to set options.}
}
\value{
Options for task queue
}
\description{
Set or get options for my package
}
\section{Supported options}{

\itemize{
  \item \code{host}: Host of PostgreSQL database
  \item \code{port}: Port of PostgreSQL database
  \item \code{user}: User name of PostgreSQL database
  \item \code{password}: Password of PostgreSQL database
  \item \code{database}: Database name of PostgreSQL database
}
}

