% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{tq_apply}
\alias{tq_apply}
\title{Apply a function with task queue}
\usage{
tq_apply(
  x,
  fun,
  project,
  resource,
  hour = 24,
  account = NULL,
  working_dir = getwd(),
  ...
)
}
\arguments{
\item{x}{A numeric vector for parallel}

\item{fun}{A function}

\item{project}{Project name}

\item{resource}{Resource name}

\item{hour}{Maximum runtime in cluster}

\item{account}{Optional. Account for cluster}

\item{working_dir}{Working directory in cluster}

\item{...}{Other arguments for fun}
}
\value{
No return values
}
\description{
Apply a function with task queue
}
