% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker}
\alias{worker}
\title{Execute a job on cluster}
\usage{
worker(project, fun, ...)
}
\arguments{
\item{project}{project name}

\item{fun}{function to run actual works which will take task id as the first argument}

\item{...}{other arguments passed to fun}
}
\value{
No return is expected from this function
}
\description{
A worker will listen task channel to get a new job, run this job and mark job
is finished until get a shutdown message to stop this function.
}
\examples{
\dontrun{
worker("test_project", mean)
}
}
